Add rename controller
