+++ GameStore/Controllers/GenreController.cs
@@ -1,6 +1,5 @@
 ﻿using GameStore.Domains.Domain;
 using GameStore.Services.Services;
-using System;
 using System.Collections.Generic;
 using System.Web.Http;
 
@@ -17,12 +16,5 @@ namespace GameStore.Controllers
         private readonly IGenreService genreService;
 
         public IEnumerable<GenreModel> GetGenres() => genreService.GetAll();
-
-        [Route("genregames/{id}")]
-        public ICollection<GameRateTransferModel> GetGenreGames(Guid id)
-        {
-            var genre = genreService.GetGenreInfoTransferById(id);
-            return genre != null ? genre.Games : new List<GameRateTransferModel>();
-        }
     }
 }
