+++ GameStore/GameStore.csproj
@@ -205,10 +205,10 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Common\ClientConfig.cs" />
-    <Compile Include="Controllers\GameController.cs" />
+    <Compile Include="Controllers\MainController.cs" />
     <Compile Include="Controllers\GenreController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
-    <Compile Include="Controllers\MainController.cs" />
+    <Compile Include="Controllers\GameController.cs" />
     <Compile Include="Controllers\ProducerController.cs" />
     <Compile Include="Controllers\StudioController.cs" />
     <Compile Include="Global.asax.cs">
